touch up qualitative test suite + add tags
