+++ chat-server/src/llmQualitativeTests/TestCase.ts
@@ -1,0 +1,11 @@
+export interface TestCase {
+  name: string;
+  expectation: string;
+  messages: TestCaseMessage[];
+  tags: string[];
+}
+
+export interface TestCaseMessage {
+  role: "user" | "assistant";
+  content: string;
+}
