+++ chat-server/src/llmQualitativeTests/getTestCasesFromYaml.ts
@@ -1,7 +1,7 @@
-import { TestCase } from "./generateChatTranscript";
 import fs from "fs";
 import path from "path";
 import yaml from "yaml";
+import { TestCase } from "./TestCase";
 
 export function getTestCasesFromYaml(fileName: string) {
   const yamlFile = fs.readFileSync(path.join(__dirname, fileName), "utf8");
